test: add backend API smoke tests and scaffolding for Docker/fuzz tests
- Add 24 backend API smoke tests using supertest (all passing)
  covering input validation, auth middleware, and response shapes
- Extract Express app setup into app.ts for testability
- Add Docker compose smoke test script (scaffolded, not yet proven)
- Scaffold Trident fuzz test harness with 14 instruction variants
  and 5 invariant checks (requires trident-cli to run)
