+++ backend/package.json
@@ -7,7 +7,8 @@
     "build": "tsc",
     "start": "node dist/server.js",
     "dev": "ts-node src/server.ts",
-    "indexer": "ts-node src/indexer.ts"
+    "indexer": "ts-node src/indexer.ts",
+    "test": "mocha --require ts-node/register tests/**/*.test.ts --timeout 30000"
   },
   "dependencies": {
     "@coral-xyz/anchor": "^0.31.1",
@@ -17,8 +18,14 @@
     "winston": "^3.11.0"
   },
   "devDependencies": {
+    "@types/chai": "^4.3.11",
     "@types/express": "^4.17.21",
+    "@types/mocha": "^10.0.6",
     "@types/node": "^20.0.0",
+    "@types/supertest": "^6.0.2",
+    "chai": "^4.3.10",
+    "mocha": "^10.2.0",
+    "supertest": "^6.3.3",
     "ts-node": "^10.9.2",
     "typescript": "^5.7.3"
   }
