+++ trident-tests/fuzz_tests/fuzz_0/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "fuzz_0"
+version = "0.1.0"
+edition = "2021"
+publish = false
+
+[dependencies]
+# Trident fuzz testing framework
+trident-client = "0.7.0"
+
+# The program under test
+sss-token = { path = "../../../programs/sss-token", features = ["no-entrypoint"] }
+
+# Anchor / Solana dependencies (must match program versions)
+anchor-lang = "0.31.1"
+anchor-spl = { version = "0.31.1", features = ["token", "token_2022", "associated_token"] }
+solana-sdk = "2.1"
+spl-token-2022 = "6.0.0"
+borsh = "0.10"
+arbitrary = { version = "1.3", features = ["derive"] }
+
+[[bin]]
+name = "fuzz_0"
+path = "test_fuzz.rs"
